// back/src/common/constants/message.enum.ts
export const enum MessageEnum {
  AUTHENTICATED = 'Autenticación correcta!',
  NO_PERMISSION = 'No tiene permiso para acceder al servicio.',
  UNAUTHORIZED = 'No autenticado, las credenciales expiraron o no se proporcionaron.',
  UNAUTHORIZED_AGETIC = 'No autenticado con Ciudadania Digital, las credenciales expiraron o no se proporcionaron.',
  FORBIDDEN = 'No autorizado, las credenciales no tienen permisos para la acción solicitada.',
  UNPROCESSABLE = 'Error al procesar la solicitud.',
  BAD_REQUEST = 'No se logró completar la operación.',
  UNSAVED_FILE = 'Error al guardar el archivo.',
  NOT_EXIST_FILE = 'No existe el archivo solicitado.',
  INVALID_FILE = 'Error de validación (archivo esperado).',
  INVALID_PARAMS = 'Parámetros no permitidos.',
  INVALID_USER = 'El usuario no existe!',
  EXIST_USER = 'El usuario ya existe!',
  INVALID_PASSWORD = 'Contraseña incorrecta del usuario.',
  INVALID_CREDENTIAL = 'Usuario o contraseña incorrectos.',
  URL_GENERADA = 'Generación de URL exitósamente.',

  OK = 'La solicitud se ha procesado correctamente.',
  ENTITY_SELECT = 'Datos devueltos correctamente.',
  ENTITY_SELECT_EMPTY = 'No se encontraron registros.',
  ERROR_CREATE = 'Error: Los datos no fueron almacenados correctamente.',
  ERROR_UPDATE = 'Error: Los datos no fueron modificados correctamente.',
  ERROR_DELETE = 'Error: Los datos no fueron eliminados (lógico) correctamente.',
  ERROR_PAGINATION = 'Error: Alguno(s) de lo(s) parámetro(s) de búsqueda/ordenación no existe(n).',
  ERROR_PRECONDITION_FAILED = 'Error: Los datos no cumplen una condición específica.',
  ERROR_PYTHON_PDF = 'Error: Ocurrio un problema con PY, contactese con el administrador!.',

  CREATED = 'El registro fue creado correctamente.',
  UPDATED = 'El registro fue modificado correctamente.',
  ENABLED = 'El estado de Habilitado fue modificado correctamente.',
  DELETED = 'El registro fue dado de baja correctamente.',
  EXIST = 'El registro ya existe!',
  NOT_EXIST = 'El registro no existe!',

  FECHA_NO_VALIDA = 'Fecha no válida; debe enviar una fecha en formato iso',
  FECHA_INICIO_NO_VALIDA = 'Fecha inicio no válida; debe enviar una fecha en formato yyyy-mm-dd.',
  FECHA_FIN_NO_VALIDA = 'Fecha fin no válida; debe enviar una fecha en formato yyyy-mm-dd.',
  FECHA_NO_VALIDA_FECHA_ACTUAL = 'La fecha no debe superar la fecha actual',
  DEFAULT = 'Sin mensaje',

  EXIST_PERSONA = 'Persona ya existe.',

  SEGIP_ESTADO_SE_ENCONTRO_REGISTRO = 'Se encontró un registro',
  SEGIP_ESTADO_NO_SE_ENCONTRO_REGISTRO = 'No se encontró el registro.',
  SEGIP_ESTADO_SE_ENCONTRO_MAS_DE_UN_REGISTRO = 'Se encontró más de un registro.',
  SEGIP_ESTADO_NO_SE_REALIZO_CONSULTA = 'No se realizó la consulta interinstitucional.',
  SEGIP_ESTADO_REGISTRO_CON_OBSERVACION = 'Registro con observación.',

  ERROR_SEGIP = 'Ocurrio un problema al consumir servicio de SEGIP',
}